+++ website/docusaurus.config.js
@@ -1,39 +1,46 @@
 // @ts-check
 // Note: type annotations allow type checking and IDEs autocompletion
 
-const lightCodeTheme = require('prism-react-renderer/themes/github');
-const darkCodeTheme = require('prism-react-renderer/themes/dracula');
+const lightCodeTheme = require("prism-react-renderer/themes/github");
+const darkCodeTheme = require("prism-react-renderer/themes/dracula");
 
 /** @type {import('@docusaurus/types').Config} */
 const config = {
-  title: 'Flutter Facebook Auth',
-  tagline: 'The easiest way to add facebook login to your flutter app, get user information, profile picture and more. Web support included.',
-  url: 'https://facebook.meedu.app',
-  baseUrl: '/',
-  onBrokenLinks: 'throw',
-  onBrokenMarkdownLinks: 'warn',
-  favicon: 'img/favicon.ico',
-  organizationName: 'darwin-morocho', // Usually your GitHub org/user name.
-  projectName: 'flutter-facebook-auth', // Usually your repo name.
+  title: "Flutter Facebook Auth",
+  tagline:
+    "The easiest way to add facebook login to your flutter app, get user information, profile picture and more. Web support included.",
+  url: "https://facebook.meedu.app",
+  baseUrl: "/",
+  onBrokenLinks: "throw",
+  onBrokenMarkdownLinks: "warn",
+  favicon: "img/favicon.ico",
+  organizationName: "darwin-morocho", // Usually your GitHub org/user name.
+  projectName: "flutter-facebook-auth", // Usually your repo name.
 
   presets: [
     [
-      '@docusaurus/preset-classic',
+      "@docusaurus/preset-classic",
       /** @type {import('@docusaurus/preset-classic').Options} */
       ({
         docs: {
-          sidebarPath: require.resolve('./sidebars.js'),
+          sidebarPath: require.resolve("./sidebars.js"),
           // Please change this to your repo.
-          editUrl: 'https://github.com/facebook/docusaurus/edit/main/website/',
+          editUrl: "https://github.com/darwin-morocho/flutter-facebook-auth/tree/master/website",
+          lastVersion: "current",
+          versions: {
+            current: {
+              label: "4.0.0",
+              path: "4.x.x",
+            },
+            "3.x.x": {
+              label: "3.x.x",
+              path: "3.x.x",
+            },
+          },
         },
-        blog: {
-          showReadingTime: true,
-          // Please change this to your repo.
-          editUrl:
-            'https://github.com/facebook/docusaurus/edit/main/website/blog/',
-        },
+
         theme: {
-          customCss: require.resolve('./src/css/custom.css'),
+          customCss: require.resolve("./src/css/custom.css"),
         },
       }),
     ],
@@ -43,19 +50,24 @@ const config = {
     /** @type {import('@docusaurus/preset-classic').ThemeConfig} */
     ({
       navbar: {
-        title: 'flutter_facebook_auth',
-       
+        title: "flutter_facebook_auth",
+
         items: [
           {
-            type: 'doc',
-            docId: 'intro',
-            position: 'left',
-            label: 'Tutorial',
+            type: "doc",
+            docId: "intro",
+            position: "left",
+            label: "Tutorial",
           },
           {
-            href: 'https://github.com/darwin-morocho/flutter-facebook-auth',
-            label: 'GitHub',
-            position: 'right',
+            href: "https://github.com/darwin-morocho/flutter-facebook-auth",
+            label: "GitHub",
+            position: "right",
+          },
+          {
+            type: 'docsVersionDropdown',
+            position: 'left',
+            dropdownActiveClassDisabled: true,
           },
         ],
       },